Re-organizing
I will edit the /etc/nginx/sites-available so that it reflects the same
structure. In order for h5ai to work, it has to sit in the web server.
So I will make the web server be the root /, so drone.nulldivision.com
would come up blank, but I'll set a 301 redirect so it goes to the /www/
directly.