Uploaded Pics and Vids
The /files/ folder lives exclusively on the server (not synced with
GitHub). This is because there will be a lot of pictures and videos
until we can sort them later.

So if you plan on adding any big file sizes, just use FTP to manually
upload them for now. When we have a finalized pictures and videos, we'll
sync it with git.